Move code to also work in embedded (ex: rand function that produces same results)
