+++ software/pc-emul/pc-emul.mk
@@ -15,6 +15,9 @@ DEFINE+= -DPC
 CPP_FILES = $(wildcard $(VERSAT_PC_EMUL)/*.cpp)
 CPP_OBJ = $(patsubst $(VERSAT_PC_EMUL)/%.cpp,$(BUILD_DIR)/%.o,$(CPP_FILES))
 
+CPP_FILES += $(VERSAT_DIR)/software/utilsCommon.cpp
+CPP_OBJ += $(BUILD_DIR)/utilsCommon.o
+
 #Units to verilate
 VERILATE_UNIT := Reg Mem Muladd VRead VWrite PipelineRegister Mux2 Merge Const Delay SwapEndian
 UNIT_VERILOG := $(foreach unit,$(VERILATE_UNIT),$(VERSAT_DIR)/hardware/src/$(unit).v)
@@ -25,6 +28,7 @@ TYPE_INFO_HDR = $(VERSAT_PC_EMUL)/versatPrivate.hpp $(VERSAT_SW_DIR)/utils.hpp $
 
 TOOL_COMMON_SRC += $(VERSAT_DIR)/software/pc-emul/parser.cpp
 TOOL_COMMON_SRC += $(VERSAT_DIR)/software/pc-emul/utils.cpp
+TOOL_COMMON_SRC += $(VERSAT_DIR)/software/utilsCommon.cpp
 TOOL_COMMON_SRC += $(VERSAT_DIR)/software/pc-emul/memory.cpp
 TOOL_COMMON_SRC += $(VERSAT_DIR)/software/pc-emul/logger.cpp
 TOOL_SRC += $(TOOL_COMMON_SRC)
@@ -60,6 +64,10 @@ $(BUILD_DIR)/typeInfo.inc: $(BUILD_DIR)/structParser.out $(TYPE_INFO_HDR)
 
 $(BUILD_DIR)/verilogWrapper.inc: $(BUILD_DIR)/verilogParser.out  $(VERSAT_SW_DIR)/pc-emul/verilogParser.cpp $(VERSAT_TEMPLATE_DIR)/unit_verilog_data.tpl
 	$(BUILD_DIR)/verilogParser.out $(BUILD_DIR)/verilogWrapper.inc -I $(VERSAT_DIR)/submodules/INTERCON/hardware/include/ -I $(VERSAT_DIR)/hardware/include/ -I $(VERSAT_DIR)/hardware/src/ $(UNIT_VERILOG)
+
+$(BUILD_DIR)/%.o: $(VERSAT_DIR)/software/%.cpp $(HDR) $(UNIT_HDR) $(VERSAT_HDR) $(BUILD_DIR)/typeInfo.inc $(BUILD_DIR)/verilogWrapper.inc
+	mkdir -p $(BUILD_DIR)
+	g++ -std=c++11 -DPC -c -o $@ $(GLOBAL_CFLAGS) $< -I $(VERSAT_SW_DIR) -I $(VERSAT_PC_EMUL) -I $(VERILATOR_INCLUDE) -I $(BUILD_DIR)/
 
 $(BUILD_DIR)/%.o: $(VERSAT_PC_EMUL)/%.cpp $(HDR) $(UNIT_HDR) $(VERSAT_HDR) $(BUILD_DIR)/typeInfo.inc $(BUILD_DIR)/verilogWrapper.inc
 	mkdir -p $(BUILD_DIR)
